+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 resolver = "2"
 members = [
+    "client",
     "dynamic-proxy",
     "plane/plane-tests",
     "plane/plane-dynamic",
