+++ client/src/sse.rs
